Add files via upload
I can't figure out how to make the elements act as inline elements. I have another version of the document which was expected to make the various companies appear side by side but it also didn't work. If someone else can figure this out or wants to discuss it this week I am open to it.